+++ default.js
@@ -129,6 +129,7 @@ module.exports = {
       'location': 'anywhere'
     } ],
     'no-with': 'error',
+    'prefer-promise-reject-errors': 'error',
     'radix': 'error',
     'require-await': 'error',
     'vars-on-top': 'error',
@@ -252,6 +253,7 @@ module.exports = {
     'no-lonely-if': 'error',
     'no-mixed-operators': 'error',
     'no-mixed-spaces-and-tabs': 'error',
+    'no-multi-assign': 'error',
     'no-multiple-empty-lines': [ 'error', { 'max': 2 } ],
     'no-negated-condition': 'error',
     'no-nested-ternary': 'error',