+++ CHANGELOG.md
@@ -11,7 +11,7 @@ details.
 
 ---
 
-## [Unreleased]
+## [1.4.0] - 2017-01-21
 ### Added
 - All published rules as of ESLint 3.14.0
 - Source to check released rules of `eslint-react-plugin` via unpkg.com.